Move helper classes to package
