+++ app/src/androidTest/java/com/example/vybe/Helpers/RecyclerViewItemCountAssertion.java
@@ -1,0 +1,29 @@
+package com.example.vybe.Helpers;
+
+import android.view.View;
+
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.test.espresso.NoMatchingViewException;
+import androidx.test.espresso.ViewAssertion;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+
+public class RecyclerViewItemCountAssertion implements ViewAssertion {
+    private final int expectedCount;
+
+    public RecyclerViewItemCountAssertion(int expectedCount) {
+        this.expectedCount = expectedCount;
+    }
+
+    @Override
+    public void check(View view, NoMatchingViewException noViewFoundException) {
+        if (noViewFoundException != null) {
+            throw noViewFoundException;
+        }
+
+        RecyclerView recyclerView = (RecyclerView) view;
+        RecyclerView.Adapter adapter = recyclerView.getAdapter();
+        assertThat(adapter.getItemCount(), is(expectedCount));
+    }
+}
